Remove Foundation and UIKit imports where they already exist in the prefix file.
Some developers (Jon Reid comes to mind) advocate keeping your files independently buildable, regardless of the contents of the prefix file, but I think it's pretty unlikely that one would ever build without Foundation and (where it was used) UIKit in the prefix (since Xcode puts them there by default), and to I find declarations like this to be more "noisy" than useful. For libraries, it makes more sense, but for app code that is unlikely to be used outside of an environment that we control, these extra declarations seem to add no value.