Number and position of targets is under the Presentation layer's control.
There is now an Application Controller in the Presentation layer that is created and triggered by the UIKit App Delegate when the app is launched. With only one screen at the moment, there is no need to expose a scene-switching interface yet, so the Application Controller is just told what screen is being shown and is charged with creating a Presenter to manage the display.

For now, this just creates a few targets on the screen, in preparation for being able to handle touches on those targets.