Use a repointable tracker for the targeting scene that delegates to subtrackers.
The initial subtracker determines whether the touch is over a target and if it is, it starts a new tracker for the connection interaction and repoints the delegating tracker at this connection tracker. If it's not over a target, it remains uncommitted until the touch moves over a target.

This will makes the move to testing for different UI elements a little simpler, as the various types of interactions can be handled by separate smaller trackers, rather than one monolithic one.