Rename "TargetContainerView" to "TrialRenderer", after the scene it represents.
Over time, the plan is to transition to using the term "Renderer" in place of "View", as it is more generic and representative of the contents of the interface, allowing for the future possibility of different types of rendering, some of which may not be visual or dynamic at all.

And the "Trial" part represents the new, more accurate term for the activity taking place in this scene: a single "trial" of the experimental exercise.