Fix a strange Xcode project ref error by removing and re-adding a test code file.
All rename refactors (the only type tried) were failing due to a error about this particular file not being referenced in any project target. The file in question *was* actually in a target, and removing it made the error go away. Re-adding it after appears to have given the file a new internal ID, and things are working again.