Add a utility class for holding an Obj-C method selector in object form.
Method selectors in Obj-C are represented by a primitive type (SEL), which is little more than a pointer to static structure built by the compiler and embedded into the application or library code. These cannot be put into Foundation collections without first converting to NSValue or NSString form. The CTDMethodSelector class is little more, but it is more intent-revealing and provides a few useful methods, and is a place to hang other related methods.