Convert existing tests to use OCHamcrest matchers, except where not advantageous.
OCHamcrest does not appear to have any facilities for catching exceptions, so continue to use XCTest for those tests. Also, in one place there is a test-only helper method currently used for verification; that should be converted to a custom Hamcrest matcher, but it's not necessary to do that immediately, so leave those tests as is for now as well.