Switch to using an Xcode workspace, in order to fix build order issues.
Discovered that, when using parallelized builds, Xcode was not waiting for the Hamcrest target to be completely built before starting on Mockito (which depends on Hamcrest's include files to have been copied to the build directory). The issue was that a subproject (here Mockito) cannot depend on another sibling subproject (only on children), and having the main project targets depend on both put Mockito and Hamcrest in the build queue together, potentially being built in parallel. (Turning off parallel builds allowed specifying the build order, but I didn't want to do that if I didn't have to.)

Switching to a workspace allows me to specify that some Mockito target depends on some Hamcrest target, so Xcode holds the Mockito build until the Hamcrest build completes. However, this only works with link dependencies ("Link Binary with Libraries"), not "Target Dependencies" (you can't select targets in other projects there), so that means Mockito must be linked with Hamcrest, and therefore the main project must not also link against Hamcrest. This is probably more correct anyway, it's just a bit fragile when the main target also uses Hamcrest directly. (If Mockito gets removed, those link references will have to be replaced with Hamcrest's library target, not simply removed.)

It also seems that the link targets are sufficient for Xcode to work out build order of the main project's targets, so all the entries under "Target Dependencies" appear to be redundant and have been removed. (It's really not clear to me why this section appears before link dependencies, if it there is rarely a need for it.)

It may make sense to reorganize the directory structure, with workspace alone with directories for each of the projects at the top level, so as to make it clear that the workspace is the thing to be opened. I've decided to leave it as is for now, until I'm sure this workspace thing is going to "stick", and while I think about whether that's really the best organization.