Interaction objects to get Presentation Model objects or manipulators, instead of renderers.
This is the next step toward introducing the Presentation Model between the input handlers and the Native UI adapter. User input is converted and passed into Interaction handlers, which manipulate Domain Model or (as here) Presentation Model objects, which then send update notifications to the Native UI over Rendering protocols.

We still need to provide a way to go from "touchable" elements (Native UI objects implementing CTDTouchable) to the corresponding Presentation Model elements; at this point, the actual color buttons do not reach at all to touches because they the touch mapper is not delivering the expected type of object (CTDSelectable) to the Interaction.