Create PM color palette enum as interface for Native UI to use.
The Native UI module should not use and Domain Model (DM) types and the Presentation Model (PM) must not use any Native UI types, so we needed to introduce a new type to act as a bridge between the layers. The color palette provides a element for each unique color that may be referred to by the PM, without tying the PM to any color model (e.g., RGB). The Native UI module can simply include a map of color palette indexes to native colors, in whatever form is most suitable.