Have the touch mapper return an associated "actuator" object to the Interaction.
It may be that the Touch Mapper interface should be more general (just mapping from the CTDTouchable to "some related object"), but for now it seems helpful to be as specific as current conditions allow. If nothing but Interaction objects will be using Touch Mappers, it seems likely that the usual association will be to an object that can perform some manipulation on model objects (what I'm calling an "actuator" for the moment; could be a "manipulator", "controller", "twiddler", or something else along those lines), so the associated object argument is more specifically named.