Not necessary to keep a collection of the cell proxies in the cell group, at the moment.
Should a method to allow removal of cells be required, then an internal list will be necessary, but since that's never needed now, and the cells (to be useful) need to be referenced (strongly) outside of the group (e.g., in a Touch Mapper), so there's no benefit to having the Cell Group retain the cell proxies as well. Only commented out for now, until I see whether removal might be required. I'm sure the code could be worked out again, but I'd prefer to leave it in as a reminder, with a comment that explains why it is commented-out.