Remove dash from '#pragma mark's within class definitions.
The dash at the start of the pragma's message results in a horizontal line separator being inserted in the jump bar navigation for the source file, and without it the pragma label is just listed in bold text, which still stands out nicely but is less intrusive than the separators. For small class definitions (my norm), the separators just aren't contributing anything useful, and I prefer the more compact form that removing the dashes produces.

There is one place I've left the dashes, in a case where the markers are separating a supporting class from the main class. It makes more sense to add a little stronger separation, to call attention to the separate classes in the navigation list.