Use compact method style in test source file (as in most of the others).
With the exception of cases where there are fewer but more dense test methods, prefer the more compact code style where the opening brace for a method appears on the same line as the method name. This actually *improves* readability in cases where the test methods are numerous but very short. It is easier to scan through all the tests for a case and they are more likely to all fit on the display at once.

Where test methods are longer or the assertion code a little more dense, keep the extra space, especially if the whole test case still fits in under 50 lines.