Add support for array and dictionary subscripting syntax with iOS 5 SDK.
The LLVM compiler has supported the subscripting syntax since the version in Xcode 4.4/4.5 (LLVM 4.0?), but the methods it requires are not included in the iOS 5 SDK headers. Nevertheless, the methods *are* actually provided for iOS 5, in a statically-linked library called ARCLite. One just has to declare that the methods exist to appease the compiler and the linker will find them without any trouble.