Tidy up test-bundle config settings for default case (Library) and special case (App tests).
The TEST_HOST setting seems to always follow the BUNDLE_LOADER value; i.e., in the case of a Library test bundle, both settings are empty; and in the case of an App test bundle, both must be set to the path of the built application.

By setting TEST_HOST to always equal BUNDLE_LOADER, and setting BUNDLE_LOADER to an empty value as default, this covers the (for me) most common case for test bundles (testing Libraries). In the rare case where I actually find a reason to test an App bundle, all the test bundle needs to define is the BUNDLE_LOADER parameter. And, in fact, Xcode sets this automatically in the xcodeproj settings when one selects an App target for a test bundle, so any setting in an xconfig file will be ignored anyway, unless the project setting is manually deleted.