Stop building test targets for "Run" build type.
Xcode 5+ no longer supports running the unit tests as part of the main build, so the only value in building the test targets at this phase is to check that the interfaces used by the tests have not been broken. But that's often just an inconvenience when you just want to do a quick experiment, and such breakages should be quickly caught in the regular case by continually running the tests while developing.