Make convenience macros for point and selector objects available outside of tests.
The macros are useful in regular code as well as in tests (though they may make more of an impact in tests, due to the greater use of constants in fixtures and verifications). By moving the macros into the headers that define the types, they can be used anywhere, with no cost to code that does not make use of them.