Add OCHamcrest & OCMockito as subprojects to support build-order control.
The CTDTestHelpers target requires the OCHamcrest public headers to be in place (we reference the copies in the build directory, as there is no distinction between public and private headers in the OCHamcrest source code), but because it would cause linking conflicts, the CTDTestHelpers static library target cannot link to the OCHamcrest static library product, so implicit dependency finding doesn't work.

We need to make a explicit entry under CTDTestHelpers's "Target Dependencies" Build Phase, and only targets within the same project are eligible, so we have to go back to putting OCHamcrest as a subproject, rather than simply being in the same workspace. (We're doing the same thing with OCMockito, just for consistency, though we haven't yet hit the same problem with it.)