Delete the Xcode workspace; no longer required now that externals are subprojects.
There doesn't seem to be anything very useful in the workspace; it tracked the repositories of the Hamcrest and Mockito git submodules, but it seems like these paths are picked up automatically when opening the project directly, so it doesn't seem like there's anything lost in ditching the explicit workspace.

Further, the xccheckout file in the automatically-generated workspace within the project folder contains only the main project element, which could change if one is working from a clone, and so really shouldn't be saved with the project. Thus, I will continue to leave this auto-workspace folder in the .gitignore list.