Use macros and better naming to improve interface to test-spy message matcher functions.
"didReceive" was a bit too generic; it is now "receivedMessage", and "didNotReceive" is now "didNotReceiveMessage". They are no longer as closely matching, but the new form feels more natural, while also reducing the risk of future naming conflicts.

Also, introducing macros as wrappers (which can be undef-ed and renamed if a conflict arises) allows for making the count-matcher argument optional (the default is "at least once", which is what is most commonly wanted) and omitting the "@selector()" part, making for more concise assertions.