Simplify touch distribution in untestable UIBridge layer --- only one touch responder per view controller.
Rather than having add and remove methods for Touch Responders, just have a single responder, set as a property of the view controller. If it's necessary or useful to have multiple responders at the application level, a multiplexing Touch Responder can be implemented there, where it can be more easily tested.