Create color palette class in rendering layer and use strings for color keys.
Rather than using an `enum` type, and be forced to wrap values in `NSNumbers` to use them as dictionary keys, use strings for the palette indices. This may make comparisons slightly slower (though probably not by much, when wrapping and unwrapping of primitives get factored in),  but allows the general UI layer to be completely unaware of any application palette keys, and adds a built-in debugging/logging description for colors.

Secondly, wrap the dictionary up in a class to represent the color palette and provide type-safe access to the palette values.