Manually trigger view controller to load its views.
Discovered, when rearranging some code, that if the view controller had not been placed in a window and made visible prior to using its properties as arguments for creating the Trial Presenter, the properties would not be initialized. This is because they are being initialized in `viewDidLoad`, which is not invoked until the views are required (e.g., when the root view becomes visible).

We add an explicit read of the view controller's `view` property to trigger this loading, and delay making the window visible until after creating the Presenter, to prove that this arrangement works, before splitting it up.