Move CTDApplication from Presentation to Application module.
Its job is to assemble the platform-independent parts of scenes, and it has to pull from several other modules: Presentation, Interaction, and Model. Thus, it should really sit outside of all of them.

Having made this move, we can now combine it with the Scene Builder to also construct the UIBridge and UIPlugins components, and wire up the whole arrangement (yet to be done, but will probably come as the application structure further evolves).