Remove unnecessary Toolbar concept and simplify color cell views in UI Bridge code.
By adding properties for the selected and unselected colors, we can lay out and configure all the color cells in the toolbar, with less code. If there is a desire to make the toolbar layout more dynamic in future, and iOS auto-layout is not available (prior to iOS 6), a UIView that just does the layout part of the former CTDUIKitToolbar class can be added back in.