Move DotView layer creation code into class methods of view class.
These were previous either part of DotView's init method or in private functions, but putting them in class methods makes it possible for a subclass to override these, or for other code to make use of the code to create similar layers. It's unlikely that either will happen within this app, but it doesn't do any harm (invoking the methods has no effect on instances of the DotView class), and is more an experiment to see how the pattern holds up.