Have VC set a connection view's frame, instead of having the view do it itself.
Previously, a connection view was automatically taking the size of its parent view when it was made a subview, but this may be too restrictive and potentially "surprising" for clients. We're only creating these views in one place anyway, so it's no harder to have the client set the frame, which is the expected way of doing things, and more flexible.