DotView Adapter should hold a weak reference to the DotView instance it wraps.
The "owner" of a view is generally the parent view, and there is rarely a need to have additional strong references to a UIView. It probably does little harm to have a strong reference, since the likelihood of retain cycles in this case is minimal, however, the only reason I used a strong ref was to avoid the extra verbiage required to avoid compiler warnings when using weak references directly. It is a bit noisier this way, but in a simple class like this and the ConnectionView adapter, the code is still clear and easy-to-read, so I think it's better to do things "by the book".