Rename CTDUIKitConnectionView -> CTDUIKitLineView.
There's nothing really application-specific about this view now; it is an example of the generic widgets that I am aiming for. The View Adapter takes care of the application-specific mapping of the Connection Rendering protocol to the generic line-drawing view.

The dot view has also been "widget-ized", so move the UIView classes into a separate group to help make this separation clear.