Use global coordinates for touch positions, instead of referencing some view.
This makes it slightly easier to convert to a subview's coordinate space as it is no longer necessary to supply a reference view in several places in order to perform the necessary conversions. One can assume that any touch-related protocol is working with global coordinates and pass `nil` for the "from" view when using UIView's conversion methods.