Define a "Connect Scene" protocol for the UIBridge to implement.
This enables the removal of both View Controller class headers from the UIBridge interface, as the UIBridge class can instantiate the appropriate class from within the module, without requiring the class to be exposed to clients, which access the object through the client-defined interface.

Currently, the Connect Scene view controller implements this directly, but in future, we may move the view controller to the Widgets module and add an adapter to UIBridge.