Make VC factory method an instance method and simplify its interface.
As an instance method, it can get all of the data it needs for configuring the View Controller from itself, so there's no need to pass any arguments to the method.

I debate this a lot, as I don't like having private methods that may unintentionally overridden, but I hope to eventually change the way the color palette value is obtained by views or their adapters, so I'll revisit this soon.