Pull CTDConnectionActivityTests test spy classes into separate source unit.
Although these are very similar to the similarly (or identically) named classes in the CTDInteractionTests module, it does not seem useful to merge the classes because the Interaction module will eventually be losing all dependencies on the Rendering interfaces, so all of the unique requirements of the test helpers in that module will disappear. Everything expected to be of use has already been pulled over, and further work will happen on the CTDApplicationTests versions.