Support separate highlighting and selection properties in Selection Editor.
There is no longer any need to treat it like a database transaction; all changes are applied immediately. There is still a need to arbitrate between multiple simultaneous touch processors (not yet being done), so the Editor model still seems useful, but the Activity or Presentation Model generally needs to be updated immediately for input actions, so it may be more a matter of exposing the appropriate properties to the Interactor, than acting like a databse.