Make a CTDDotPair a proper value object type, with copying, equality and hash support.
Copying was added to support code that ended up getting replaced, but it's bound to be useful sooner or later, so it doesn't make sense to throw it away. The rest was added at the same time as copying support, when I noticed it missing; again, it's not currently being used, but I think a DotPair is really a value type, so it should follow the conventions.