Add dot coordinate and distance to per-step results.
The distance, in particular, may be useful to see whether longer trial
times were due to a sequence with unusually long distances between points,
taking longer to make the connections. But dot position may also be useful
as it may indicate that certain parts of the screen are more challenging
to reach than others.