Display a pop-up before a trial, with an instructional message and a "Begin" button.
The message will eventually be used to tell the user whether the next
trial is a practice or recorded trial, and what number it is. There is
also a less obvious "Exit" button that allows the experimenter to return
to the start menu if, for some reason, the participant cannot complete
the block of trials.