Change all UIColor properties from 'copy' to 'strong', for comp. w/iOS 5
Prior to iOS 6, UIColor does not support the NSCopying protocol, so any
attempt to copy a UIColor instance results in a crash (despite the
compiler never complaining, even when using the iOS 5 headers).

Since we're in control of all of the code that calls our classes, I have
chosen to change the types of the UIColor properties to 'strong', for
expediency, rather than trying to patch in NSCopying support for iOS 5.
This would be possible, but it doesn't really add any safety in this case,
as it takes a concerted effort to break UIColor's immutability, and we're
mostly just passing around values set in Interface Builder, and there's
no need to modify them.