Color selection interactors should only clear changes they previously made.
In other words, do not call the Selection Editor's 'clearHighlighting' or
'clearSelection' methods unless that specific interaction previously (and
actively) set a highlighted element or made a selection.

This fixes the problem with the touch-and-hold interactor where a second
touch (to make a dot connection) cancels the first touch's color selection.