Reduce the number of trials to 10 (plus 2 practice).
The 20 pairs per trial feels about right (it could be reduced, but I want
to be sure the trials are long enough to show a difference in time between
the conditions), but at about 35-50 seconds per trial, 12 trials will take
about 10 minutes, which is about as long as I would want anyone to have to
go. I may further shave this down to 2+8 if piloting indicates that a
plateau is reached earlier and participants are showing signs of fatigue
or boredom.