Save application state before each trial so it can be restored after a crash.
This can also be used to reverse an accidental "Exit" from a trial block,
as the new state is not written until a new configuration is chosen and
the first new trial is started. Just exit the app and kill the process;
when it restarts, it will read the saved state and return to the pre-Exit
position.