Also reload the Trial Block Results data when restoring application state.
We save the trial durations with the application state, and then "replay"
those durations into the new results object when the state is restored.
This will overwrite the data in the block results file, but the new data
should be identical to the old data. The individual trial results files
are not touched (except that any in-progress trial will be overwritten
when the trial is restarted, but that is as intended).