Add time taken to perform connection part of each step to trial results.
This is recorded from the first touch on the starting dot, after the
corresponding color has been selected in the palette. It isn't possible to
determine the exact moment where the user performs a context switch in
his or her mind, but this gives us a _low-end_ conservative figure for
the connection time, which can be subtracted from the total step time to
further isolate the difference in performing the color-selection part of
the task.