In modal style, move palette to P-hand side (e.g., right for righties).
It was observed that requiring users to reach across the screen to select
a color may take slightly longer, especially because "homing" one's hand
over the palette after a line would block the view of the next dot; thus,
it was observed that this might negatively affect the single-handed UI
times.

The first 6 participants used the UI with the palette on the left; I will
try another 6 the opposite way and compare results. Unless times get
worse, or the results are mixed, I will continue with this layout for the
remainder of the trials.